[Front] migration of Position.js to tsx (#2843)
